+++ src/main/java/com/xhadl/yournotion/Repository/SurveyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SurveyRepository extends JpaRepository<SurveyEntity, Long> {
+public interface SurveyRepository extends JpaRepository<SurveyEntity, Integer> {
     public SurveyEntity findById(int id);
     public Page<SurveyEntity> findAllByOrderByIdDesc(Pageable pageable);
     public SurveyEntity save(SurveyEntity surveyEntity);
