+++ src/main/java/com/xhadl/yournotion/Repository/SurveyWantRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SurveyWantRepository extends JpaRepository<SurveyWantEntity, Long> {
+public interface SurveyWantRepository extends JpaRepository<SurveyWantEntity, Integer> {
     public SurveyWantEntity findByUserIdAndSurveyId(int userId, int surveyId);
     public SurveyWantEntity save(SurveyWantEntity surveyWantEntity);
 }
